services/watchdog/requirements.txt
httpx==0.27.2
pydantic==2.8.2
pyyaml==6.0.2
prometheus_client==0.20.0
PyGithub==2.3.0